refactor: abstract controller logic
